refactor: reorganise how ndk files are structured
