+++ embrace-android-sdk/src/main/cpp/anr/anr.h
