+++ embrace-android-sdk/src/main/cpp/anr/emb_anr_manager.c
