+++ embrace-android-sdk/src/main/cpp/crashmarker/file_marker.c
