+++ embrace-android-sdk/src/main/cpp/safejni/jni_util.c
