+++ embrace-android-sdk/src/main/cpp/schema/stack_frames.h
