+++ embrace-android-sdk/src/main/cpp/signals/signals_c.c
@@ -12,11 +12,11 @@
 #include <dlfcn.h>
 #include "signals_c.h"
 #include "../unwinders/unwinder.h"
-#include "../utilities.h"
-#include "../file_marker.h"
-#include "../file_writer.h"
+#include "../utils/utilities.h"
+#include "../crashmarker/file_marker.h"
+#include "../serializer/file_writer.h"
 #include "signal_utils.h"
-#include "../emb_log.h"
+#include "../utils/emb_log.h"
 #include "../utils/string_utils.h"
 
 #define EMB_SIG_HANDLER_COUNT 6
