+++ embrace-android-sdk/src/main/cpp/unwinders/unwinder.c
@@ -4,8 +4,8 @@
 
 #include "unwinder.h"
 #include "unwinder_stack.h"
-#include "../utilities.h"
-#include "../emb_log.h"
+#include "../utils/utilities.h"
+#include "../utils/emb_log.h"
 #include "../utils/string_utils.h"
 #include <dlfcn.h>
