+++ embrace-android-sdk/src/main/cpp/utils/emb_log.c
