+++ embrace-android-sdk/src/main/cpp/utils/utilities.c
