+++ embrace-android-sdk/src/test/cpp/CMakeLists.txt
@@ -2,8 +2,13 @@ include_directories(
         ${CMAKE_CURRENT_SOURCE_DIR}/include
         ../../main/cpp
         ../../main/cpp/3rdparty
+        ../../main/cpp/anr
+        ../../main/cpp/crashmarker
+        ../../main/cpp/jnibridge
         ../../main/cpp/safejni
         ../../main/cpp/sampler
+        ../../main/cpp/schema
+        ../../main/cpp/serializer
         ../../main/cpp/signals
         ../../main/cpp/unwinders
         ../../main/cpp/utils
